refactor: errorFallbackLogic 리팩토링
