+++ frontend/src/components/ErrorFallbackNotificationMessage/type.ts
@@ -1,5 +1,3 @@
 import { ErrorType } from 'types/internal';
 
 export type ErrorFallbackNotificationMessageProps = ErrorType;
-
-export type UseErrorFallbackNotificationMessageProps = ErrorType;
