+++ frontend/src/components/ErrorFallbackSubscriptionButton/type.ts
@@ -1,5 +1,3 @@
 import { ErrorType } from 'types/internal';
 
 export type ErrorFallbackSubscriptionButtonProps = ErrorType;
-
-export type UseErrorFallbackSubscriptionButtonProps = ErrorType;
