+++ frontend/src/hooks/useErrorFallbackLogic.ts
@@ -1,17 +1,14 @@
-import { UseErrorFallbackSubscriptionButtonProps } from './type';
 import { authApiClient } from 'apis/apiClient';
 import { useEffect } from 'react';
 import { useSetRecoilState } from 'recoil';
 import { isLoginState } from 'recoil/auth/atoms';
+import { ErrorType } from 'types/internal';
 
 import useSnackBar from 'hooks/useSnackBar';
 
 import { CLIENT_PATH } from 'constants/path';
 
-export const useErrorFallbackSubscriptionButton = ({
-  errorCode,
-  errorMessage,
-}: UseErrorFallbackSubscriptionButtonProps) => {
+export const useErrorFallbackLogic = ({ errorCode, errorMessage }: ErrorType) => {
   const renderSnackBar = useSnackBar();
   const setIsLogin = useSetRecoilState(isLoginState);
 
@@ -24,15 +21,15 @@ export const useErrorFallbackSubscriptionButton = ({
     });
 
     // 로그아웃 처리
-    if (errorCode === 2002 || errorCode === 2003 || errorCode === 4001) {
+    if (
+      errorCode === 2002 ||
+      errorCode === 2003 ||
+      errorCode === 4001 ||
+      errorCode === 9001 ||
+      errorCode === null
+    ) {
       authApiClient.deleteAuth();
       setIsLogin(false);
     }
   }, [errorCode, errorMessage]);
-
-  const handleClickToggleButton = () => {};
-
-  return {
-    handleClickToggleButton,
-  };
 };
